Made some modifications to Sentiment Classifier.
Used a LinkedHashMap instead of a hashmap because the elements are stored in order or iteration. Added a few print methods. Removed punctuation marks and symbols from afinn library but just for each word or phrase which makes it easier to check if equals with sorted input text. Modified constructor to only load stopwords based on language the user has chosen.